C prototipo/API/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
mysql-connector-python
python-dotenv
sqlalchemy
pydantic-settings
pymysql
cryptography